Rewrite the experiment
Gameinfo/Game.py include the data structure of
A game
	-4 player
		-Position
		-Hand
	-Double Dummy Solver Result
Gameinfo/calculate.py include a calculator data structure of
A score recorder
	-4 position DDS result
		-[Spade, Heart, Diamond, Club, NT]
	-the result at NS and WE
	-the RULE that how to calculate the score
		-HCP
		-Long
		-Short
		-Loser
		-Pavlicek
	-the score matrix of
		-HCP
		-Long(increase how many score during every length)
		-Short(increase how many score during every length)
Functions
	-listadd(list1, list2)
		Add the elements in list1 and list2 by index, so the
		lenght of result is equal to length of list1 and list2.
	-getcurr(DDS, score)
		Calculate the coefficient correlation between DDS and
		score which calculate by the score recorder.
Gameinfo/test is the data set for check the calculator correct or not.
Gameinfo/DDSresult.txt is the FULL data set.

Experiment_Corr.py is the Executable file that calculate the coefficient
correlation of data set.