Change the result of longgest
Change the way to get the longgest suit of the pairs of players.
----From----
Record the win tricks and the suit which have the longgest card between
the pair of players.
If there are more than one suit have the longgest card, compare the win
tricks between them.
If there are same winning tricks, Then compare the suit.
----After----
Record every suit and tricks which have the longgest card.