Add two method to split the region of win tricks
input: the score appear times
output: the edge of each score

meanAppearTimesOfEachScore: use the mean number of scores in each win
tricks to get the edge.
middleAppearTimesOfEachScore: use the middle number of scores in each
win tricks to get the edge.

And remove the makeGapEPS function, this function should not implement over here.