Add new feature and new scorer
nScorer() is the scorer in the future, it should calculate the score
according to the formulaid. The formulaid in the future is a mask of the
combination of features.

nSet_scorematrix() is the score matrix setter in the future, like the
nScorer().

HCP_notrump() is the feature that used in formula12. It is a different
feature with HCP, that calculate the HCP by trump suit or not. The
HCP_notrump don't have that abstraction.

TrumpLong() is the feature that used in formula12. Get extra strength
when that suit is trump suit. If not trump suit, it may not a good
distribution.

Nontrumpshort() is the feature taht used in formula12. Get extra
strength when the short suit is not trump suit. If the trump is short,
it may not a good situation.