Implement the second experiment for ch4.2
We use this experiment to find the better hyper-parameters of the
combinations. The dataset we use here is the same as the previous
experiment. After we find the better hyper-parameters for the
combinations we are interested in, these hyper-parameters will be
the parameters for the investigation in the next section.